+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# 0.5.1 (June 2022)
+
+- Read multi-resolution pyramids in `ome_zarr.reader.Well`, thanks to @tcompa ([#208](https://github.com/ome/ome-zarr-py/pull/208))
+
+# 0.5.0 (June 2022)
+
+- Add `fmt` option to `write_label_metadata` and store version ([#206](https://github.com/ome/ome-zarr-py/pull/206))
+
+# 0.4.2 (May 2022)
+
+- Relax version detection ([#189](https://github.com/ome/ome-zarr-py/pull/189))
+- Fix format warning duplication ([#190](https://github.com/ome/ome-zarr-py/pull/190))
+- Fix plate pyramid loading ([#195](https://github.com/ome/ome-zarr-py/pull/195))
+
+# 0.4.1 (Mar 2022)
+
+- Unify docstrings in `ome_zarr.writer` module ([#185](https://github.com/ome/ome-zarr-py/pull/185))
+
+# 0.4.0 (Mar 2022)
+
+- Add API for writing data according to the `image-label` spec: `write_labels`, `write_multiscale_labels` & `write_label_metadata` ([#178](https://github.com/ome/ome-zarr-py/pull/178), [#184](https://github.com/ome/ome-zarr-py/pull/184))
+- Remove `byte_order` from `write_image` signature ([#184](https://github.com/ome/ome-zarr-py/pull/184))
+- Deprecate `chunks` argument in favor of `storage_options` ([#184](https://github.com/ome/ome-zarr-py/pull/184))
+
 # 0.3.0 (Feb 2022)
 
 ## Specification changes